feat. redux
redux로 로그인 상태 전역적으로 관리.
과정에서 Rootlayout과 함께 사용하지 못하는 문제를 각 폴더들의 layout을 만들고, 만든 layout 내에서 관리하도록 변경
