--- back/src/main/java/com/demo/library/security/config/WebConfig.java
@@ -1,22 +1,0 @@
-package com.demo.library.security.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-//CORS 커스텀 설정
-@Configuration
-@EnableWebMvc
-public class WebConfig implements WebMvcConfigurer {
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/**")
-                .allowedOrigins("http://localhost:3000")
-                .allowedMethods("GET", "POST","PATCH","DELETE","OPTIONS") // 허용할 HTTP method
-                .exposedHeaders("*") //header 노출
-                .allowedHeaders("*") // 요청 헤더 중 서버에서 허용하는 헤더
-                .allowCredentials(true) // 쿠키 인증 요청 허용
-                .maxAge(3000); // 원하는 시간만큼 pre-flight 리퀘스트를 캐싱
-    }
-}
