+++ back/src/main/java/com/demo/library/security/controller/AuthController.java
@@ -3,27 +3,34 @@ package com.demo.library.security.controller;
 import com.demo.library.response.ResponseCreator;
 import com.demo.library.response.dto.SingleResponseDto;
 import com.demo.library.security.entity.RefreshToken;
+import com.demo.library.security.handler.CustomLogoutHandler;
 import com.demo.library.security.jwt.dto.TokenDto;
 
 import com.demo.library.security.jwt.jwttokenizer.JWTTokenizer;
 import com.demo.library.security.oauth.google.GoogleAuthService;
 import com.demo.library.security.oauth.google.GoogleDto;
 import com.demo.library.security.oauth.google.GoogleJWTService;
+import com.demo.library.security.repository.RefreshTokenRepository;
 import com.demo.library.security.service.JWTAuthService;
 import com.demo.library.security.oauth.kakao.KaKaoDto;
 import com.demo.library.security.oauth.kakao.KakaoAuthService;
 import com.demo.library.security.oauth.kakao.KakaoJWTService;
 import com.demo.library.user.entity.User;
+import com.demo.library.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.ServletException;
 import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
 
 
 @Controller
@@ -36,6 +43,8 @@ public class AuthController {
     private final KakaoJWTService kakaoJWTService;
     private final GoogleAuthService googleAuthService;
     private final GoogleJWTService googleJWTService;
+    private final RefreshTokenRepository refreshTokenRepository;
+    private final UserRepository userRepository;
 
 
     @GetMapping("/login-form")
@@ -51,6 +60,13 @@ public class AuthController {
     @PostMapping("/login")
     public String login() {
         return "home";
+    }
+    @PostMapping("/logout")
+    public ResponseEntity<String> logout(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        new CustomLogoutHandler(refreshTokenRepository, userRepository).onLogoutSuccess(request, response, authentication);
+
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @PostMapping("/token/refresh")
