+++ back/src/main/java/com/demo/library/security/handler/CustomLogoutHandler.java
@@ -1,5 +1,6 @@
 package com.demo.library.security.handler;
 
+import com.demo.library.security.entity.RefreshToken;
 import com.demo.library.security.repository.RefreshTokenRepository;
 import com.demo.library.user.entity.User;
 import com.demo.library.user.repository.UserRepository;
@@ -7,10 +8,13 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.web.authentication.logout.LogoutSuccessHandler;
+import org.springframework.web.util.WebUtils;
 
 import javax.servlet.ServletException;
+import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.transaction.Transactional;
 import java.io.IOException;
 import java.util.Optional;
 
@@ -20,11 +24,14 @@ public class CustomLogoutHandler implements LogoutSuccessHandler {
     public final UserRepository userRepository;
     HttpStatus httpStatusToReturn = HttpStatus.OK;
     @Override
-    public void onLogoutSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException {
-        // repository 에서 refreshToken 삭제
-        if (authentication != null)
-            userRepository.findByName(authentication.getName()).ifPresent(refreshTokenRepository::deleteByUser);
-
+    @Transactional
+    public void onLogoutSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException{
+        Cookie refreshTokenCookie = WebUtils.getCookie(request, "refreshToken");
+        if(refreshTokenCookie != null) {
+            String token = refreshTokenCookie.getValue();
+            Optional<RefreshToken> refreshTokenOptional = refreshTokenRepository.findByToken(token);
+            refreshTokenOptional.ifPresent(refreshTokenRepository::delete);
+        }
         response.setStatus(this.httpStatusToReturn.value());
         response.getWriter().flush();
     }
