--- front/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "untitled2",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "@hookform/resolvers": "^3.3.4",
-    "axios": "^1.6.5",
-    "next": "14.0.4",
-    "react": "^18",
-    "react-dom": "^18",
-    "react-hook-form": "^7.49.3",
-    "react-redux": "^9.1.0",
-    "redux": "^5.0.1",
-    "zod": "^3.22.4"
-  },
-  "devDependencies": {
-    "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
-    "autoprefixer": "^10.0.1",
-    "eslint": "^8",
-    "eslint-config-next": "14.0.4",
-    "postcss": "^8",
-    "tailwindcss": "^3.3.0",
-    "typescript": "^5"
-  }
-}
